Diplom_Django/diplom_dj/requirements.txt
Django==2.2.8
Pillow==7.0.0
django-mptt
django-admin-thumbnails




